Create Menu; Larger Dungeon; Reach Flag to "win"

+++ controller/PlayerController.gd
@@ -11,15 +11,21 @@ const MARKER_COLOR := Color.GREEN
 const MARKER_COLOR_DEAD := Color.RED
 
 @export var controlledMonster: Monster
-@onready var dungeon = $".."
+@onready var dungeon: Dungeon = $Dungeon
 @onready var camera: Camera2D = $Camera2D
 @onready var marker: TextureRect = $TextureRect
+
+@onready var menuScreen: CanvasLayer = $MenuScreen
+@onready var scoringLabel: Label = $MenuScreen/ColorRect/ColorRect3/ScoringLabel
+@onready var pauseButton: Button = $HUD/ButtonPause
 
 var markerMotion := 0.0
 
 
 func _ready() -> void:
+	get_tree().paused = true
 	marker.modulate = MARKER_COLOR
+	switchControlledMonster("first")
 
 func _process(delta: float) -> void:
 	if controlledMonster:
@@ -33,6 +39,9 @@ func _process(delta: float) -> void:
 			markerMotion = sin(Time.get_unix_time_from_system() * 2*PI) * MARKER_AMPLITUDE
 		marker.position = focus + MARKER_POSITION
 		marker.position.y += markerMotion
+
+		if dungeon.flag.has_overlapping_bodies():
+			displayMenu()
 
 func _physics_process(delta: float) -> void:
 	if not is_instance_valid(controlledMonster):
@@ -54,7 +63,7 @@ func switchControlledMonster(which) -> void:
 	var newMonster := getMonster(which)
 	controlledMonster = newMonster
 	marker.modulate = MARKER_COLOR
-	if oldMonster.isDead and newMonster != oldMonster:
+	if oldMonster and oldMonster.isDead and newMonster != oldMonster:
 		oldMonster.delete()
 
 func getMonster(which) -> Monster:
@@ -70,6 +79,8 @@ func getMonster(which) -> Monster:
 			return monsters[0]  # Loop around to first monster
 		else:
 			return monsters[rank+1]
+	elif which == "first":
+		return monsters[0]
 	return null
 
 func getAllMonsters() -> Array[Monster]:
@@ -93,3 +104,26 @@ func sortByHorizontalPosition(a: Node2D, b: Node2D) -> bool:
 		return true
 	else:
 		return false
+
+
+func displayMenu() -> void:
+	menuScreen.visible = true
+	scoringLabel.text = "Game Over!\n\n%s out of your 8 Goombas were stomped" % [8 - getAllMonsters().size() + (1 if controlledMonster.isDead else 0)]
+	get_tree().paused = true
+	pauseButton.text = "Resume Game"
+
+func hideMenu() -> void:
+	menuScreen.visible = false
+	pauseButton.visible = true
+	get_tree().paused = false
+	pauseButton.text = "Pause Game"
+
+
+func _on_button_level_1_pressed():
+	hideMenu()
+
+func _on_button_pause_pressed():
+	if get_tree().paused:
+		hideMenu()
+	else:
+		displayMenu()
